+++ js/main.js
@@ -1,3 +1,5 @@
-//data.js является матрешкой. main<-data<-util(переменные); main<-data<-random-integer(функция по рандомгым числам)ж
-import './data.js';
+
+import './util.js';
 import './module7-task1.js';
+import './module8-task1.js';
+
